requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.40
aiosqlite==0.20.0
pydantic-settings==2.7.1
